requirements-webhook.txt
# WhatsApp Webhook Service Dependencies
# Production-ready dependencies for EA integration

# Core web framework
flask>=3.0.0
flask-limiter>=3.5.0
gunicorn>=21.2.0

# HTTP and API communication
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Database dependencies for EA integration
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Redis for client registry and caching
redis>=5.0.0

# Security and authentication
cryptography>=41.0.0

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Voice integration (optional)
elevenlabs>=2.14.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# AI model integration (optional)
openai>=1.0.0